changed doubles to floats and other stuff on gpu

+++ ray_trace.c
@@ -9,7 +9,7 @@
 #define WINDOW_Y 10
 #define WINDOW_MAX 10
 
-typedef struct { double x,y,z; } vec_t;
+typedef struct { float x,y,z; } vec_t;
 
 vec_t add(vec_t v, vec_t w){
     vec_t res;
@@ -19,28 +19,28 @@ vec_t add(vec_t v, vec_t w){
     return res;
 }
 
-vec_t scale(double c, vec_t v){
+vec_t scale(float c, vec_t v){
     v.x *= c;
     v.y *= c;
     v.z *= c;
     return v;
 }
 
-double dot(vec_t u, vec_t v){
+float dot(vec_t u, vec_t v){
     return (u.x * v.x) + (u.y * v.y) + (u.z * v.z);
 }
 
 vec_t direction(vec_t end, vec_t start){
     vec_t dir = add(end, scale(-1, start));
-    double norm = sqrt(dot(dir, dir));
+    float norm = sqrt(dot(dir, dir));
     return scale(1/norm, dir);
 }
 
 vec_t sample_direction(){
-    double phi, cos_th, sin_th;
+    float phi, cos_th, sin_th;
     vec_t v;
-    phi    = (double) rand() / (double) RAND_MAX * 2 * M_PI;
-    cos_th = (double) rand() / (double) RAND_MAX * 2 - 1;
+    phi    = (float) rand() / (float) RAND_MAX * 2 * M_PI;
+    cos_th = (float) rand() / (float) RAND_MAX * 2 - 1;
     sin_th = sqrt(1 - cos_th * cos_th);
 
     v.x = sin_th * cos(phi);
@@ -51,15 +51,15 @@ vec_t sample_direction(){
 
 
 typedef struct {
-    double *surface, max, y;    // Windows are always parallel to y axis since
+    float *surface, max, y;    // Windows are always parallel to y axis since
     int size;                   // the viewer is assumed to be at the origin
 } window_t;                     // facing in y direction
 
-void init_window(window_t *w, int size, double y_coordinate, double w_max){
+void init_window(window_t *w, int size, float y_coordinate, float w_max){
     w->y = y_coordinate;
     w->max = w_max;
     w->size = size;
-    w->surface = calloc(size * size, sizeof(double));
+    w->surface = calloc(size * size, sizeof(float));
     assert(w->surface);
 }
 
@@ -73,11 +73,11 @@ int on_surface(window_t *w, vec_t co_ord) {
 
 void save_surface(window_t *w,  char *file_name){
     FILE *f = fopen(file_name,"w");
-    fwrite(w->surface, sizeof(double), w->size * w->size, f);
+    fwrite(w->surface, sizeof(float), w->size * w->size, f);
     fclose(f);
 }
 
-void add_brightness(window_t *w, vec_t co_ord, double b){
+void add_brightness(window_t *w, vec_t co_ord, float b){
     int x,y;
     x = (co_ord.x + w->max) / 2 / w->max * w->size;
     y = (co_ord.z + w->max) / 2 / w->max * w->size;
@@ -86,25 +86,22 @@ void add_brightness(window_t *w, vec_t co_ord, double b){
 
 // Ray tracing simulation of a sphere at coordinates `center` and radius `radius`.
 // Assumes observer is at origin facing in y direction dowards `window`
-void serial_ray_tracing(window_t* window, vec_t center, vec_t light, double radius, int n_rays){
+void serial_ray_tracing(window_t* window, vec_t center, vec_t light, float radius, int n_rays){
 
-    double radius_sq, center_sq;
+    float radius_sq, center_sq;
     radius_sq = radius * radius;
     center_sq = dot(center, center);
 
     for(int n=0; n<n_rays; n++){
         vec_t ray, ray_window, ray_sphere, sph_dir, light_dir;
-        double brightness, t, x = -1;
+        float brightness, t, x = -1;
 
         do{
             ray = sample_direction();                     // Ray originates from the origin
             ray_window = scale(window->y / ray.y, ray); // Ray's intersection with window
-
             if(!on_surface(window, ray_window)) continue;
-
             t = dot(ray, center);                   // t and x are algebra steps needed
             x = t * t + radius_sq - center_sq;      // to solve for ray-sphere intersection
-
         } while(x < 0);
 
         ray_sphere = scale(t - sqrt(x), ray);
